Cleanup grafana_auth
